==> Pandas/02-Lesson/Video_Lecture_NBs/NB_10_Importing Data from Web.ipynb <==
# Part 2: Full Data Workflow A-Z

## Importing Data from the Web  with pd.read_html()

import pandas as pd

url = "https://en.wikipedia.org/wiki/1976_Summer_Olympics_medal_table"

pd.read_html(url)[3] # now pos3

wik_1976 = pd.read_html(url)[3] # now pos3

wik_1976.head()

wik_1976.tail()

wik_1976.info()

url2 ="https://en.wikipedia.org/wiki/1996_Summer_Olympics_medal_table"

pd.read_html(url2)[3] # now pos3

wik_1996 = pd.read_html(url2)[3] # now pos3

wik_1996.head()

wik_1996.info()

wik_1976.to_csv("wik_1976.csv", index= False)
wik_1996.to_csv("wik_1996.csv", index= False)
