==> Pandas/02-Lesson/Video_Lecture_NBs/NB_14_Reshaping DataFrames.ipynb <==
# Part 2: Full Data Workflow A-Z

## Reshaping DataFrames

### Transposing DataFrames

import pandas as pd

titanic = pd.read_csv("titanic.csv")

titanic.head()

titanic.T

titanic.transpose()

titanic.info()

titanic = titanic.transpose()

titanic.head()

titanic.info()

titanic = titanic.T

titanic.head()

titanic.info()

titanic.survived.mean()

#titanic.groupby(["sex", "pclass"]).survived.mean()

titanic = pd.read_csv("titanic.csv")

titanic.info()

titanic.groupby(["sex", "pclass"]).survived.mean().unstack().T

titanic.T.T.info()

### Pivoting DataFrames with pivot()

import pandas as pd

table1 = pd.read_csv("table1.csv")

table1.head(10)

table1.tail(10)

table1.info()

table1.shape

table1.pivot(index = "Country", columns = "Medal", values = "Count")

table1_piv = table1.pivot(index = "Country", columns = "Medal", values = "Count").fillna(0)

table1_piv

table1_piv.shape

table1_piv.info()

table1.head()

table1.set_index(["Country", "Medal"]).unstack(fill_value = 0)

table1_piv

### Limits of pivot()

import pandas as pd

table2 = pd.read_csv("table2.csv")

table2.head(20)

table2.tail(10)

table2.info()

table2.shape

#table2.pivot(index = "Country", columns = "Medal", values = "Count")

table2.groupby(["Country", "Year", "Medal"]).Count.sum().unstack().fillna(0)

table2_piv = table2.groupby(["Country", "Year", "Medal"]).Count.sum().unstack().fillna(0)

table2_piv

table2_piv.info()

table2_piv.shape

### pivot_table()

import pandas as pd

table2 = pd.read_csv("table2.csv")

table2.head(10)

table2.pivot_table(index = "Country", columns = "Medal", values = "Count", aggfunc="sum", fill_value= 0, margins= True, margins_name="Total")

table2.groupby(["Country", "Medal"]).Count.sum().unstack(fill_value = 0)

agg = table2.groupby(["Country", "Medal"]).Count.sum().unstack(fill_value = 0)

agg.head()

agg["All"] = agg.sum(axis = 1)

agg.loc["All"] = agg.sum(axis = 0)

agg

### pd.crosstab()

import pandas as pd

titanic = pd.read_csv("titanic.csv")

titanic.head()

titanic.info()

pd.crosstab(titanic.sex, titanic.pclass, margins=True, normalize= True)

titanic.groupby(["sex", "pclass"]).pclass.count().unstack()

pd.crosstab(index = titanic.sex, columns =  titanic.pclass, values=titanic.age, aggfunc="mean")

titanic.pivot_table(index = "sex", columns = "pclass", values = "age", aggfunc = "mean")

titanic.groupby(["sex", "pclass"]).age.mean().unstack()

pd.crosstab(index = titanic.sex, columns =  titanic.pclass, values = titanic.fare, aggfunc = "sum", margins=True, normalize= True)

summer = pd.read_csv("summer.csv")

summer.head()

summer.info()

pd.crosstab(index = [summer.Year, summer.Country], columns = summer.Medal, aggfunc = "count",
            values = summer.Athlete, margins = True).fillna(0)

summer.pivot_table(index=["Year", "Country"], columns = "Medal", aggfunc = "count",
                   values = "Athlete", fill_value=0, margins = True)

summer.groupby(["Year", "Country", "Medal"]).Medal.count().unstack(fill_value = 0)

### Melting DataFrames with melt()

import pandas as pd

table_2012 = pd.read_csv("table_2012.csv")

table_2012

table_2012.shape

table_2012.melt(id_vars = "Country", value_vars = ["Gold", "Silver", "Bronze"], var_name="Medal", value_name="Count")

melt_2012 = table_2012.melt(id_vars = "Country", value_vars = ["Gold", "Silver", "Bronze"], 
                            var_name = "Medal", value_name = "Count")

melt_2012

melt_2012.pivot(index = "Country", columns = "Medal", values = "Count").sort_values("Gold", ascending = False)
